Remove fast path in the core pick_next_task function as this breaks the implementation.
As the fast path assumes that if there is no other task in the system
or if only CFS tasks are currently running, there is no to ask other
scheduling classes whether they want to run anything, running the idle
thread on some cores did not work properly. Now whit this fast path
removed this problem is solved.