Fix bugs in the CPU acquire and release methods as well as in the redistribute function.
The local CPU acquire was done too late in the process which resulted in
delays in the execution of the process.

The energy task redistribution function assumed that if a task's thread
gets dequeued it is ok to continue running other tasks on the
corresponding CPU. However, this assumption is not fully true. In some
circumstances it is necessary to completely give up the CPU instead.