Rework the rescheduling and blocking of CPUs.
These functions also use smp_single_call_async in combination with an
internal data structure so that they happen as fast as possible and not
asynchronously as it happen until now. Additionally if between a send
request and its processing a local decision is done, the corresponding
request is not executed. Another important part is, that only one
request can be in the air at one point in time.