Implement conditional wait-for-update.
Before, every time we made an update of the RAPL counters, we waited for
them to settle. While this is necessary for short running programs, this
is rather useless if one has long-running CPU-bound applications that
don't yield the CPU often. The procfs now also reports this new
information and also got a new interface that is easier to parse via a
program (similar to stat vs status).